Add user PDO handling
The InOut data structure is now extended to support the user PDOs and
the send and receive example now reads and writes these additional PDO
values. This patch will only work correctly with a SII/EEPROM
configuration which supports the necessary space in the SyncManager
configuration.

To make the operation more clean the provided ESI configuration file for
this device and the object dictionary in module_ethercat has to be
extended too.