Add optional time measurement
This optional code measures the time (in nano seconds) between consecutive
calls to pdo_handle_ecat(). This funciton is called periodically due to a timer
(-> setitimer()) which triggers SIG_ALRM. If active the time difference to the
last call to this funciton and the mean call time is printed to stdout.

The time measurement is avtivated by setting TIME_MEASUREMENT to 1 in the
Makefile.