Change to new SDO/CoE API
To signal any change in the OD the API for the CoE access had to be
changed to a interface handling in module_ethercat. The addition here
respects these change.

Because of conflicts with the ethercat service API in the cia402-wrapper
all the PDO handling is temporarily removed from this test app.