Add new module_pdo_handler
This module is used to handle the PDO communication for the application.
This is the next step in the separation of PDO and SDO communication for
the EtherCAT client.

Now the dependencies for the simple PDO test application is
significantly reduced since only the pdo handler and the ethercat
service is necessary. The list of used modules is reduced significantly.