Restructure the example usage for the end config signal
The "configuration_ready()" notification is raised every time the
master set the slave in operation mode and is about to start the
cyclic/realtime operation.

As soon as the master sets operation mode it is assumed the
configuration of all necessary objects in the object dictionary is
finished and the values are correctly transfered from the master to the
slave (which the master checks with a sequence of upload and download
requests). The client (e.g. the motor control) is responsible to update
the configuration from the object dictionary everytime the notification
is raised.