Add missing state and bitmask for the first states
One state was missing, but it is necessary to announce the fault
handling is currently underway.

The status word bitmask for the states FAULT_REACTION_ACTIVE and FAULT
is necessary because some bits have to be '0' while other bits are don't
care. In order to keep the don't care bits all bits in the mask are
zeroed and after that the necessary ones are set.