Refactor of drive action according to state
The update of the local variables and the update of the PDO to send is
done right before further drive actions. Quickstop handling is further
consolidated and the use of the requested operation mode is also
straightend.

Currently everything is CSP specific and has to be extended to support
other modes of operation. Please refer to IEC 61800-7 for further
information on operation mode specific behavior.