Refactor state variables
Variable decalration and initialization is mostly done the C99 way, this
shortens this section is the code. Unused variables are moved out of the
way in a commented out section which can be removed after
reconstruction. Some variables are still unused but present in the code
because the further usage is currently unclear.