Fix timeout recognition
If packets are received again after packets are once dropped the
comm_inactive_flag is reset. So we have to receive no PDOs for more than
one second to really trigger the timeout and go to Fault state.

The inital fault state, if the communication comes back, is
automagically reset by the TwinCAT master.