Change opmode handling
Due to some misunderstandings about the PDO object 0x6060 (Op Mode) we
now read the opmode from the object dictionary and set this as default
opmode. If a opmode request occures, the slave checks if the request is
in the supported opmodes. If yes, then the new opmode is set in the
state "Switch on Disabled". If the requested opmode is not in the
supported modes we will stick to the currently active opmode.