Add quickstop abort if velocity is less than 200 rpm
This prevents the overshoot of the quickstop profiler if the motor is
nearly stopped but the steps to perform are not zero. The profiler tries
to drive the motor to it's new position with the result, the motor
accelerates again (mostly in the other direction) until all the steps
are performed.