Add command line arguments
Namely the ncurses display can be turned on by '-w' switch, this slows down the
processing but gives a neat overview of what is send and what is received.

With the '-d' switch some information are printed before the processing starts.
Currently only the current domain registration is shown. This can be helpfull
during debugging.