Rework of pdo and sdo handling
With the PDO values for controlword, tuning_command and user_miso the
value of various objects can be requested via separate channel.

The controlword contains the current command for reading (= 1) or
writing (= 2). The index and subindex of the object is encoded in
user_mosi, in user_miso the result, in case of read is stored. In
torque_command the value to write is delivered. The statusword holds the
reusult of the last action (0 for success, 1 in case of error).