Remove continus mode option from visibility
Since this mode only returns usefull information if app_demo_slave_sdo_handling
is running on the slave but this master demo can be used with all other slave
firmatwares to request the online values of the object dictionary this option
can lead to confusion on the user side. So this is removed for now.