Proper use of the Polarity 0x606C CiA402 object
The effect is exactly like described in the standard:
When the position polarity bit is set the position command will be inverted in CSP mode.
When the velocity polarity bit is set the velocity command will be inverted in CSV mode.

When the either position or velocity commands are inverted, all the feedbback (position/velocity/torque will be inverted).
Attention, the effect is only active in CSP and CSV opmodes, when the opmode is not CSP or CSV the polarity switches back to normal and the feedback will not be inverted anymore.
So changing the mode can actually change the feedback direction if we change from an opmode where the polarity is different.
The master needs to take this into account.