Add conditional linkage of ethercat lib
If the build configuration is Debug than the standalone version of the
ethercat library is used. Otherwise the standard version for use on
devices with a bootloader is used. This makes it easy to develop and
test with xrun the application, system integration or deployment happens
only with the release build of the firmware.