Add proper access rights handling
I still don't see the advantage of anonymous enum in this particular
case. Makeing it named one can use it as return value (as intended) and
allow the compiler for proper type checking.

The check of the access rights is done in the ethercat service CoE
handling with the use of the od_get_access() interface method which
proofes pretty usefull here.