Fix used modules and used libraries
The module_ethercat was renamed lib_ethercat, the standalone build of
the library is intended for use during devleopment (= Debug build) and
currently it is also used for the nnt patched/ old COM_ECAT modules.
The default build of the library is intended to be used with the next
revision of COM_ECAT of a patched one.