Add flash write feature to command structure
There are serious memory issues on the tile where the object dictionary
is stored, so the command service needs to run on a different tile.

To be able to restore the object dictionary on the device to a
previously saved state a additional command to read the object is
defined.