Add new dictionary structure
To separarate the dictionary declaration from the access functions I
declared the dictionary variables extern to store them in a separate
source file. The dictionary variables could become quit large so the
access function could be hidden behind all this noise, furthermore I
want to separate the acccess function (which shouldn't change) from the
generated content.

The SDO access is separated from the dictionary and is primary intended
to allow access from the EtherCAT side. The functions
`sdo_entry_{get,set}_value()` can also be used by the application to
retrieve and set values for specific parameters.

The object dictionary contains a subset of the objects and corresponding
entries with values and strings. Default, min and max are not yet
implemented. The code for the dictionary will be generated.