Add value access functions
The access function handle the master request flag. If the master writes
a value the flag is set, a reading by the local application clears the
flag again. With this mechanism it is possible for the application to
see and read recently updated values by the master.

The write and read function check in case a request from the master
occures if the master has sufficient rights to read or write the entries
value. Error number is set appropreately and holds the last error
occured. The behaviour of the error variable is similar to the behaviour
of the errno variable in the C standard library.