Add implementation for object and entry description
This is a pita, if the `struct _sdoinfo_entry_description` is in
`co_interface.h` I cannot include inside the C context because of this
interface definition. But in `sdo.h` it is actually not really well
suited since the ethercat service also needs knowledge of this struct
(recently the struct is replicated along with the whole co_interface
header.

FIXME Need to clean this up.