Add command object handling to new sdo api
Also added defines for direction of request for future improvement.

The currenth interface does not distinguish if the sdo request comes
from the master or from the app. The dictionary is able to handle both
request, it just needs the information from where it comes.

Future work (next task): is to make to interfaces for the
canopen_interface_service. One for the communication with the COM side,
and one for the communication with the APP side.