Add valueinfo functionality
The parameter `valueinfo` is a bitfield used to decide which additional
information are put into the entry description. This could be:
- unit information (if available, 0 otherwise)
- default value
- minimum value (if available)
- maximum value (if available)

The current challenge is that default, min and max have arbitrary sizes
which are not known in advance and all four information could be part of
the entry reply.

**Attention:**
Since the Etherlab master seem to miss this feature this code is not
very well tested.