Add request for change flags
These interface calls are added to access the chaged flags of entries in
the object dictionary. The change flag changes when the master initiates
a SDO upload request, the flag is cleared when the application reads the
entries value.

With these interface calls it is possible to get the number of entries
which has changed, get the index and subindex of one of the last changed
entries and check a entry given by index and subindex if it has changed
since the last time the value was read by the application.