Use the correct format of HexDecValue type for object info units
In ETG.2000 Table 78: Content description of ObjectInfoType the data
type of Unit element is HexDecValue. The description for HexDecValue in
Table 6 says: represents a hexadecimal value either in hexadecimal or
decimal format, so 12345 is decimal and #x12345 is hexadecimal. In
accordance to that this commit removes 0 units and adds #x prefix to
units that didn’t have one.