Use float for PID coefficients SDO parameters.
This means the master has to send the value formatted as a float otherwise it will not be interpreted correctly.
The local PID coefficients are still in integer so they are cast from float in the config manager.
The csv config parser now support float parsing (using the '.' character).
The sdo write function is app_master_cyclic and tuning will read the SDO type and send the value as float or integer accordingly.