Replace default values for padding subitems from 07000040 to 07000000
The default value of 07000040 for padding subitems was done by hand to
silence the complains of the CTT about mismatched index in mapped
objects.

The index for padding subitems of PDO objects should be #x0000.

The source code generator for co_dictionary.c will automatically set
the next unused object index from manufacturer area to padding entries
in order to avoid the problem with IgH EtherCAT Master that does not
handle the padding index #x0000 correctly.