Fix problem with hang of the ethercat communication
As described in Wrike: https://www.wrike.com/open.htm?id=174640271 the
ethercat communication came to a halt if in opmode the storing was
triggered. This patch fixes the hang in the check for changed objects
which only occured if the command object (currently at index 0x2000) was
also changed. Now everything should be fine.