Fix Object dictionary initialization
Now the config from all the services is loaded first in the dictionary:
Then it try to read the csv config file on flash.
And if it fail it reloads the config from all the services.

So in case the config file exists but is incomplete the missing objects are still loaded from the services.