Remove torque control interface from network drive
Now set/get config and reset fault for the torque control are done through
the motion control interface.
Also removed the call to set brake when opmode is not set because in this case the motorcontrol is already disabled.

This is done to simplify the network drive service and remove unnecessary call to the torque controller.