Add reading and setting of RTC value
The RTC is set by object 0x2D02 by the master. If the device is in PREOP
EtherCAT state the change of the variable is recognised and the RTC
registers are set accordingly.

Since the data time TIME_OF_DAY used as CiA object data type describes
the time relative to 1.1.1984 the calculation from TIME_OF_DAY to
localtime is done in the functions located in cia_time.xc. The
milliseconds in the RTC register are not set because it is most likely
that the read number of microseconds significantly differ from the real
ones.

While touching the code also fixed the interface variable name for
i2c_master.