Manually fix padding PDO mappings
For some reasons unknown in the moment the mapping of the padding
objects is again 0x0000. According to the EtherCAT convention this
is perfectly okay, but the IgH master has problems handling these
mapped paddings correctly. The PDO ordering is messed up and the
values cannot be read correctly.

To avoid this problem the index of the paddings is set to 0x4000.