Fix read, store and memory consumption of device_alias
Due to memory shortage the file service now only has one buffer for FoE
operation and other string operation.

The storage and reading of the device name string in the file is fixed
by so the string is recognised as string (with trailing '\0'). Please
keep in mind that the content of the object does not necessarily have
the trailing '\0'!